Fix for issue #4
`command` doesn't always exist. As @alexpbrown pointed out in his pull
request, it's a built-in fuction of bash and not always available within
Ruby backticks.

However, there's really no need to try to determine the path to
`man2html` in the first place, cause if the path can be determined, it
logically should already be available to you for execution.

Additionally, Manservant::Server always sets the `:man2html_path` option
to the bundled binary. The option exists on Manservant::ManPage simply
to keep the API to the class sane.